separated router out of startUp.ts
